+++ src/main/java/curso/gestionAcademia/util/Secuencial.java
@@ -4,19 +4,19 @@ package curso.gestionAcademia.util;
  * Secuencial: Lista de secuenciales
  */
 public class Secuencial {
-    private static long secuencialAsignatura = 0;
-    private static long secuencialProfesor = 0;
-    private static long secuencialAula = 0;
-    private static long secuencialAlumno = 0;
+    private static Integer secuencialAsignatura = 0;
+    private static Integer secuencialProfesor = 0;
+    private static Integer secuencialAula = 0;
+    private static Integer secuencialAlumno = 0;
 
-    public static long getSecuencialAsignatura(String tipo) {
+    public static Integer getSecuencialAsignatura(String tipo) {
         String codSig = (String) Constantes.COD_ASIG.get(tipo) + (++Secuencial.secuencialAsignatura);
         Integer secAsigAux = new Integer(codSig);
         
-        return secAsigAux.longValue();
+        return secAsigAux;
     }
 
-    public static long getSecuencialProfesor() {
+    public static Integer getSecuencialProfesor() {
         return ++Secuencial.secuencialProfesor;
     }
 
@@ -24,7 +24,7 @@ public class Secuencial {
         return Constantes.AULA + (++Secuencial.secuencialAula);
     }
 
-    public static long getSecuencialAlumno() {
+    public static Integer getSecuencialAlumno() {
         return ++Secuencial.secuencialAlumno;
     }
 }
